# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Hospital Navigation API
  description: Hệ thống điều hướng và quản lý bệnh viện thông minh
  version: 1.0.0
servers:
  - url: http://group3.it4788.sukkaito.id.vn
    description: Production
  - url: http://localhost:8080
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: OK
        data:
          type: object
          nullable: true
...
paths:
...
  /api/map/get_nodes:
    get:
      tags: [Map]
      summary: Danh sách POI trên bản đồ
      parameters:
        - name: map_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/get_edges:
    get:
      tags: [Map]
      summary: Danh sách cạnh (adjacency)
      description: |
        Tính edges tự động từ grid_data (4 hướng N/S/E/W).
        Trả về danh sách cạnh giữa các ô walkable liền kề, mỗi cạnh chỉ 1 chiều (from_location < to_location).
      parameters:
        - name: map_id
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK — `data.edges[]` là mảng `{from_row, from_col, from_location, to_row, to_col, to_location}`, `data.total` là tổng số cạnh
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/get_meta:
    get:
      tags: [Map]
      summary: Metadata bản đồ
      parameters:
        - name: map_id
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/get_depts:
    get:
      tags: [Map]
      summary: Danh sách khoa/phòng
      parameters:
        - name: node_type
          in: query
          schema: { type: string }
        - name: ward_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/search_location:
    get:
      tags: [Map]
      summary: Tìm kiếm vị trí
      parameters:
        - name: keyword
          in: query
          required: true
          schema: { type: string }
        - name: map_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/get_landmarks:
    get:
      tags: [Map]
      summary: Danh sách landmark
      parameters:
        - name: map_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/map/sync_full:
    get:
      tags: [Map]
      summary: Đồng bộ toàn bộ bản đồ
      parameters:
        - name: map_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }

  # Legacy aliases used by external Jest test suite.
...
  /api/admin/get_maps:
    get:
      tags: [Admin-Map]
      summary: Lấy danh sách map
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/admin/set_active_map:
    post:
      tags: [Admin-Map]
      summary: Đổi Map Active
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [map_id]
              properties:
                map_id: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/admin/upload_map:
    post:
      tags: [Admin-Map]
      summary: Tải lên file .map
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                map_name: { type: string }
                rows: { type: string }
                cols: { type: string }
                file:
                  {
                    type: string,
                    format: binary,
                    description: "File .map từ MovingAI",
                  }
                image_file:
                  {
                    type: string,
                    format: binary,
                    description: "Ảnh preview của map (định dạng PNG/JPG/WEBP)",
                  }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/admin/upload_output:
    post:
      tags: [Admin-Map]
      summary: Tải lên file output.json (MAPF)
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                map_id: { type: integer }
                file: { type: string, format: binary }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
  /api/admin/export_map:
    get:
      tags: [Admin-Map]
      summary: Xuất file .map
      security: [{ BearerAuth: [] }]
      parameters:
        - name: filename
          in: query
          required: true
          schema: { type: string, example: "warehouse_small.map" }
          description: Tên file .map cần xuất (lấy từ trường map_file_path của map)
      responses:
        "200":
          description: OK (Trả về file .map binary)
  /api/admin/delete_map:
    delete:
      tags: [Admin-Map]
      summary: Xóa map theo ID (không cho xóa map đang active)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [map_id]
              properties:
                map_id: { type: integer, description: "ID của map cần xóa" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              { schema: { $ref: "#/components/schemas/Response" } }
  /api/admin/deactivate_map:
    post:
      tags: [Admin-Map]
      summary: Tắt trạng thái active của map (set is_active = false)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [map_id]
              properties:
                map_id: { type: integer, description: "ID của map cần deactivate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              { schema: { $ref: "#/components/schemas/Response" } }
...
  /api/admin/edit_node:
    post:
      tags: [Admin-Map]
      summary: Sửa Node
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
                x: { type: integer }
                y: { type: integer }
                name: { type: string }
                type: { type: string }
                is_landmark: { type: boolean }
                wheelchair_accessible: { type: boolean }
                is_accessible: { type: boolean }
                capacity: { type: integer }
                details: { type: string }
                open_hours: { type: string }
                custom_weight: { type: number }
      responses:
        "200":
          description: |
            1000 — Cập nhật thành công.
            2001 — Thiếu id.
            4001 — Node không tồn tại.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Response" }
...
  /api/sos/get_detail:
    get:
      {
        tags: [SOS],
        summary: Chi tiết SOS,
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/sos/get_list:
    get:
      {
        tags: [SOS],
        summary: Danh sách SOS (Staff),
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/sos/respond:
    post:
      {
        tags: [SOS],
        summary: Nhận SOS (Staff),
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/sos/resolve:
    post:
      {
        tags: [SOS],
        summary: Đóng SOS (Staff),
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }

  # ============================================================
  # CHAT
  # ============================================================
...
  /api/engine/solve:
    post:
      {
        tags: [Engine],
        summary: Chạy Dijkstra,
        security: [{ BearerAuth: [] }],
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        required: [start_location, dest_location, mode_id],
                        properties:
                          {
                            start_location: { type: integer },
                            dest_location: { type: integer },
                            mode_id: { type: string },
                          },
                      },
                  },
              },
          },
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  # /api/engine/update_cost — ĐÃ LOẠI BỎ
  # Weight được tính tự động từ flow density, manual override gây xung đột với simulation.
  /api/engine/convergence:
    get:
      {
        tags: [Engine],
        summary: Trạng thái hội tụ,
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/engine/set_params:
    post:
      {
        tags: [Engine],
        summary: Thiết lập tham số,
        security: [{ BearerAuth: [] }],
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        properties:
                          {
                            max_agents: { type: integer },
                            time_step_ms: { type: integer },
                            cost_multiplier: { type: number },
                          },
                      },
                  },
              },
          },
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
...
  /api/engine/clear_cache:
    post:
      {
        tags: [Engine],
        summary: Xóa cache Dijkstra,
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/engine/load_mapf:
    post:
      {
        tags: [Engine],
        summary: Load MAPF output,
        security: [{ BearerAuth: [] }],
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        required: [file_path],
                        properties: { file_path: { type: string } },
                      },
                  },
              },
          },
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/engine/mapf_positions:
    get:
      {
        tags: [Engine],
        summary: Vị trí agents tại timestep,
        security: [{ BearerAuth: [] }],
        parameters:
          [
            {
              name: timestep,
              in: query,
              schema: { type: integer, default: 0 },
            },
          ],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }
  /api/engine/mapf_info:
    get:
      {
        tags: [Engine],
        summary: MAPF metadata,
        security: [{ BearerAuth: [] }],
        responses:
          {
            "200":
              {
                description: OK,
                content:
                  {
                    application/json:
                      { schema: { $ref: "#/components/schemas/Response" } },
                  },
              },
          },
      }

  # ============================================================
  # MISSING APIS (Appended automatically)
  # ============================================================
...